apy/requirements.txt
Flask
flask-swagger-ui
flask-cors
numpy
panda
matplotlib
torch
torchvision
torchaudio
pillow